addressed review comments

+++ core/src/main/java/org/apache/spark/shuffle/sort/ShuffleInMemorySorter.java
@@ -176,7 +176,7 @@ final class ShuffleInMemorySorter {
     int offset = 0;
     if (useRadixSort) {
       offset = RadixSort.sort(
-        array, (long)pos,
+        array, pos,
         PackedRecordPointer.PARTITION_ID_START_BYTE_INDEX,
         PackedRecordPointer.PARTITION_ID_END_BYTE_INDEX, false, false);
     } else {
